REFACTORED: policy datamodel REFACTORED: HomeAdapter & PolicyTransformer
