+++ app/src/main/java/com/blockchain/transactionsChallenge/MainActivity.kt
@@ -3,9 +3,7 @@ package com.blockchain.transactionsChallenge
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
 import androidx.fragment.app.Fragment
-import com.blockchain.cuvva.HomeFragment
-import com.blockchain.transaction.ui.transactionsFragmentCoroutine.TransactionsFragmentCoroutine
-import com.blockchain.transaction.ui.transactionsFragmentRx.TransactionsFragmentRx
+import com.blockchain.cuvva.ui.home.HomeFragment
 import dagger.android.AndroidInjection
 import dagger.android.AndroidInjector
 import dagger.android.DispatchingAndroidInjector
