+++ core/src/main/java/com/blockchain/core/network/policy/datamodel/Documents.kt
@@ -1,0 +1,6 @@
+package com.blockchain.core.network.policy.datamodel
+
+data class Documents(
+    val certificate_url: String,
+    val terms_url: String
+)
