--- core/src/main/java/com/blockchain/core/network/policy/datamodel/PolicyResponse.kt
@@ -1,5 +1,0 @@
-package com.blockchain.core.network.policy.datamodel
-
-data class PolicyResponse(
-    val policies: List<Policy>
-)
