+++ core/src/main/java/com/blockchain/core/network/policy/datamodel/Vehicle.kt
@@ -1,0 +1,10 @@
+package com.blockchain.core.network.policy.datamodel
+
+data class Vehicle (
+    val vrm: String,
+    val prettyVrm: String,
+    val make: String,
+    val model: String,
+    val variant: String,
+    val color: String
+)
