+++ cuvva/src/main/java/com/blockchain/cuvva/ui/home/policyTransformer/CarTypes.kt
@@ -1,0 +1,9 @@
+package com.blockchain.cuvva.ui.home.policyTransformer
+
+enum class CarTypes{
+    VOLKSWAGEN,
+    MINI,
+    MERCEDES_BENZ,
+    NISSAN,
+    FORD
+}
