+++ cuvva/src/main/java/com/blockchain/cuvva/ui/home/policyTransformer/PolicyTypes.kt
@@ -1,0 +1,7 @@
+package com.blockchain.cuvva.ui.home.policyTransformer
+
+enum class PolicyTypes{
+    POLICY_CREATED,
+    POLICY_FINANCIAL_TRANSACTION,
+    UNKNOWN
+}
