+++ cuvva/src/main/res/layout/itemview_home.xml
@@ -74,6 +74,7 @@
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintTop_toBottomOf="@id/itemview_home_reg_plate"
             app:layout_constraintHorizontal_chainStyle="packed"
+            android:visibility="gone"
             tools:cpb_progress="90" />
 
         <TextView
@@ -85,6 +86,7 @@
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintStart_toEndOf="@id/itemview_home_policy_time_remaining_percent"
             app:layout_constraintTop_toTopOf="@id/itemview_home_policy_time_remaining_percent"
+            android:visibility="gone"
             tools:text="55 minutes remaining" />
 
     </androidx.constraintlayout.widget.ConstraintLayout>
